pyproject.toml
[project]
name = "dagflow"
version = "0.9.0"
description = "The DAGFlow package is a python implementation of the dataflow programming with the lazy graph evaluation."
authors = [{ name = "DAGFlow Team", email = "dagflow@jinr.ru" }]
maintainers = [{ name = "DAGFlow Team", email = "dagflow@jinr.ru" }]
license = "MIT"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "contextlib2",
  "h5py",
  "latexdatax",
  "matplotlib",
  "numba",
  "numpy",
  "ordered-set",
  "pandas",
  "pygraphviz",
  "pyyaml",
  "schema",
  "scipy",
  "tabulate",
  "uproot",
  "plotille",
  "nestedmapping",
]

[project.urls]
homepage = "https://github.com/dagflow-team/dagflow"
repository = "https://github.com/dagflow-team/dagflow"
documentation = "https://github.com/dagflow-team/dagflow/wiki"
"Bug Tracker" = "https://github.com/dagflow-team/dagflow/issues"

[project.optional-dependencies]
dev = ["pytest", "pytest-cov", "coverage"]

[build-system]
requires = ["setuptools", "wheel", "packaging"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
exclude = [
  "tests*",
  "docs*",
  "example*",
] # Исключает целые папки при необходимости
